+++ tests/DotEnvTestCase.php
@@ -124,7 +124,9 @@ abstract class DotEnvTestCase extends TestCase
     protected function makeDir(): string
     {
         $dir = sys_get_temp_dir() . '/dotenv-test-' . uniqid('', true);
-        mkdir($dir, 0777, true);
+        if (!mkdir($dir, 0777, true) && !is_dir($dir)) {
+            self::fail(\sprintf('Could not create temp directory "%s".', $dir));
+        }
         $this->tempDirs[] = $dir;
 
         return $dir;